refactor: re-organize codebase
