+++ front-end/src/components/Toolbar/Toolbar.css
@@ -7,7 +7,7 @@
     display: flex;
     justify-content: space-between;
     align-items: center;
-    z-index: 5;
+    background: black;
   }
   
   .logo {
@@ -44,3 +44,4 @@
   .navbar a:hover::before {
     width: 100%;
   }
+
