+++ front-end/src/components/home/LineAnimationComponent.tsx
