+++ front-end/src/pages/home/HomePage.tsx
@@ -3,7 +3,6 @@ import { useNavigate } from 'react-router-dom';
 import { motion } from 'framer-motion';
 import { LineAnimationComponent } from '../../components/home/LineAnimationComponent';
 import PersonComponent from '../../components/home/PersonComponent';
-import Toolbar from '../../components/Toolbar/Toolbar';
 // import HeadphonesBarAfter from "../../assets/images/HomePage/headphone-bar-after.svg";
 
 import './HomePage.css';
@@ -20,7 +19,7 @@ function HomePage() {
           setAnimate(false);
           setHeadClicked(true);
           setTimeout(() => {
-            navigate('/SongsPage');
+            navigate('/input/songs');
           }, 1000);
       }, 1000);
     }
@@ -32,41 +31,42 @@ function HomePage() {
   
   
     return (
-        <body>
-          {/* //TODO: fix this animate/ exit issue */}
-          <motion.div exit={{opacity : 0}} transition={transition} className="main-container">
-            <Toolbar />
+      <>
+        {/* //TODO: fix this animate/ exit issue */}
+        <div style={{ background: "black" }}>
+          <motion.div exit={{ opacity: 0 }} transition={transition} className="main-container">
             <div className="filler-container"></div>
             <LineAnimationComponent startAnimation={animate} />
             <div className="person-container">
               {/* //TODO: same issue with this one */}
-                <motion.div key="text-container" exit={{opacity : 0}} transition={transition} className="text-container">
-                  <div className="text-bottom">
-                    this is a super cool website
-                    <br></br>
-                    blah blah blah blah blah balh blah blah blah 
-                    <br></br>
-                    yeet
-                  </div>
-                  {/* <div className="text-bottom">
+              <motion.div key="text-container" exit={{ opacity: 0 }} transition={transition} className="text-container">
+                <div className="text-bottom">
+                  this is a super cool website
+                  <br></br>
+                  blah blah blah blah blah balh blah blah blah
+                  <br></br>
+                  yeet
+                </div>
+                {/* <div className="text-bottom">
                     a different bit of text to 
                     <br></br>
                     make this website even cooler lolllllllllll 
                     <br></br>
                     yeehaw
                   </div> */}
-                </motion.div>
+              </motion.div>
               <motion.div
                 initial={{ scale: 1 }}
                 transition={{ duration: 0.75 }}
                 animate={animateStyle}
                 className="expanded-container"
               >
-              <PersonComponent handleHeadClick={handleHeadClick} headClicked={headClicked}/>
+                <PersonComponent handleHeadClick={handleHeadClick} headClicked={headClicked} />
               </motion.div>
             </div>
           </motion.div>
-        </body>
+        </div>
+      </>
     );
   }
