+++ front-end/src/pages/input/GenerationPage.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function GenerationPage() {
+  return (
+    <div style={{ color: "white"}}>GenerationPage</div>
+  )
+}
