+++ front-end/src/pages/result/GeneratedPlaylistPage.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function GeneratedPlaylistPage() {
+  return (
+    <div>GeneratedPlaylistPage</div>
+  )
+}
